requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==1.2.0          # via flask-restful
click==6.7                # via flask
flask-restful==0.3.5
flask-sqlalchemy==2.2
flask==0.12
itsdangerous==0.24        # via flask
jinja2==2.9.5             # via flask
markupsafe==1.0           # via jinja2
python-dateutil==2.6.0    # via aniso8601
pytz==2016.10             # via flask-restful
pyyaml==3.12
six==1.10.0               # via flask-restful, python-dateutil
sqlalchemy==1.1.6         # via flask-sqlalchemy
werkzeug==0.12.1          # via flask
